+++ gui/brusheditor.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 # This file is part of MyPaint.
-# Copyright (C) 2014-2018 by the MyPaint Develoment Team.
+# Copyright (C) 2014-2018 by the MyPaint Development Team.
 # Copyright (C) 2007-2013 by Martin Renold <martinxyz@gmx.ch>
 #
 # This program is free software; you can redistribute it and/or modify