+++ lib/gichecks.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2018 by the MyPaint Develoment Team.
+# Copyright (C) 2018 by the MyPaint Development Team.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by