+++ mypaint.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python2
 # This file is part of MyPaint.
 # Copyright (C) 2007-2013 by Martin Renold <martinxyz@gmx.ch>
-# Copyright (C) 2013-2018 by the MyPaint Develoment Team.
+# Copyright (C) 2013-2018 by the MyPaint Development Team.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -238,7 +238,7 @@ def get_paths():
 
 
 def init_gettext(localepath):
-    """Intialize locales and gettext.
+    """Initialize locales and gettext.
 
     This must be done before importing any translated python modules
     (to get global strings translated, especially brushsettings.py).